pyproject.toml
[project]
name = "claude-md-transcripts"
version = "0.1.0"
description = "Convert Claude Code session JSONL transcripts to markdown and index them in qmd."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "David Sanders" }]
dependencies = [
    "pydantic>=2.7",
    "python-slugify>=8.0",
    "click>=8.1",
]

[project.scripts]
claude-md-transcripts = "claude_md_transcripts.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["claude_md_transcripts"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "ty>=0.0.1a1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "D"]
ignore = ["D1", "D200", "D212"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.coverage.run]
branch = true
source = ["claude_md_transcripts"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80
